+++ dojo-blog/src/views/Home.vue
@@ -10,6 +10,7 @@
 
 <script>
 import { ref } from 'vue'
+import getPosts from '../composables/getPosts'
 
 // component imports
 import PostList from '../components/PostList.vue'
@@ -18,23 +19,7 @@ export default {
   name: 'Home',
   components: { PostList },
   setup() { 
-    const posts = ref([])
-    const error = ref(null)
-
-    const load = async () => {
-      try {
-        let data = await fetch('http://localhost:3000/posts')
-        if(!data.ok) {
-          throw Error('no available data')
-        }
-        posts.value = await data.json()
-        console.log(posts.value)
-      }
-      catch(err) {
-        error.value = err.message
-        console.log(error.value)
-      }
-    }
+    const { posts, error, load } = getPosts()
 
     load()
